requirements.txt
Flask==3.0.2
python-docx==1.1.0
python-pptx==0.6.23
reportlab==4.1.0
